+++ src/hssm/__init__.py
@@ -16,7 +16,7 @@ from .datasets import load_data
 from .defaults import show_defaults
 from .hssm import HSSM
 from .link import Link
-from .param_refactor import UserParam as Param
+from .param import UserParam as Param
 from .prior import Prior
 from .simulator import simulate_data
 from .utils import set_floatX
