+++ src/hssm/defaults.py
@@ -24,7 +24,7 @@ from .likelihoods.analytical import (
     logp_lba3,
 )
 from .likelihoods.blackbox import logp_ddm_bbox, logp_ddm_sdv_bbox, logp_full_ddm
-from .param_refactor.utils import _make_default_prior
+from .param.utils import _make_default_prior
 
 LogLik = Union[str, PathLike, Callable, Op, type[Distribution]]
 ParamSpec = Union[float, dict[str, Any], bmb.Prior, None]
