+++ tests/param/test_regression_param.py
@@ -7,8 +7,8 @@ import hssm
 from hssm import Prior
 from hssm.defaults import default_model_config
 from hssm.link import Link
-from hssm.param_refactor import UserParam
-from hssm.param_refactor.regression_param import RegressionParam, _make_priors_recursive
+from hssm.param import UserParam
+from hssm.param.regression_param import RegressionParam, _make_priors_recursive
 from hssm.prior import HSSM_SETTINGS_DISTRIBUTIONS, HDDM_SETTINGS_GROUP
